retry screen when not all services are available
