+++ ts/sagas/startup/loadServiceRequestHandler.ts
@@ -1,22 +1,10 @@
-import * as pot from "italia-ts-commons/lib/pot";
 import { readableReport } from "italia-ts-commons/lib/reporters";
 import { call, Effect, put, select } from "redux-saga/effects";
 import { ActionType } from "typesafe-actions";
 import { BackendClient } from "../../api/backend";
 import { contentServiceLoad } from "../../store/actions/content";
 import { firstServicesLoad, loadService } from "../../store/actions/services";
-import {
-  isFirstVisibleServiceLoadCompletedSelector,
-  isVisibleServicesContentLoadCompletedSelector
-} from "../../store/reducers/entities/services/firstServicesLoading";
-import {
-  servicesByIdSelector,
-  ServicesByIdState
-} from "../../store/reducers/entities/services/servicesById";
-import {
-  visibleServicesSelector,
-  VisibleServicesState
-} from "../../store/reducers/entities/services/visibleServices";
+import { isAnyServicesContentLoadFailedSelector } from "../../store/reducers/entities/services";
 import { SagaCallReturnType } from "../../types/utils";
 /**
  * A generator to load the service details from the Backend
@@ -51,39 +39,17 @@ export function* loadServiceRequestHandler(
     }
   } catch {
     yield put(loadService.failure(action.payload));
+  }
 
-    // If at least one service content loading fails, the first services load is considered as failed
-    const visibleServices: VisibleServicesState = yield select(
-      visibleServicesSelector
+  // If at least one service loading fails, the first services load is considered as failed
+  const isAnyServicesContentLoadFailed = yield select(
+    isAnyServicesContentLoadFailedSelector
+  );
+  if (isAnyServicesContentLoadFailed) {
+    yield put(
+      firstServicesLoad.failure(
+        Error("Failed to load the content of one or more services")
+      )
     );
-    const isFirstServiceLoading: pot.Pot<boolean, Error> = yield select(
-      isFirstVisibleServiceLoadCompletedSelector
-    );
-    const isVisibleServicesContentLoadCompleted = yield select(
-      isVisibleServicesContentLoadCompletedSelector
-    );
-
-    if (
-      pot.isNone(isFirstServiceLoading) &&
-      pot.isSome(visibleServices) &&
-      isVisibleServicesContentLoadCompleted
-    ) {
-      const servicesById: ServicesByIdState = yield select(
-        servicesByIdSelector
-      );
-      const isServicesContentLoadFailed =
-        visibleServices.value.findIndex(service => {
-          const serviceContent = servicesById[service.service_id];
-          return serviceContent !== undefined && pot.isError(serviceContent);
-        }) !== -1;
-
-      if (isServicesContentLoadFailed) {
-        yield put(
-          firstServicesLoad.failure(
-            Error("Failed to load the content of one or more services")
-          )
-        );
-      }
-    }
   }
 }
